Create validator
This is a pseudo "init" script, which allows Validator to be run in a shared environment that is not tied to one user's SSH session.  WARNING: this shared web server approach does not have any sort of access control around it; anyone who can connect to the server and knows the port would have access.  The script allows you to start, stop and check status of the Validator REST server.  To automatically start and stop Validator on boot and shutdown, this script should be placed in /etc/init.d/, and respective "K**" and "S**" symlinks should be created in the desired runlevel directory (e.g. /etc/init.d/rc3.d).  The WD variable must be set to the path of the Validator directory.